Refactor CI process to boost it by avoiding full reinstall all the time
Fresh install of node_modules is just too slow. Since we are doing 3 of
them (statinamic, boilerplate - test, docs - rendering), we need to
cache node_modules…

The process is a bit weird but it has been made this way to ensure “npm
prune” (which is a crucial step when you use travis node_modules cache
to check that we don’t have deps not listed in package.json (which will
create “untestable” issue with fresh install)) will not delete some
deps that it should not “in real life” (because statinamic is linked,
but not the entire folder directly otherwise we got too many
node_modules (dev deps) in
test-boilerplate/node_modules/statinamic/node_modules.
Similar workaround is made for docs.