`@phenomic/core`: Prevent watcher from crashing when a file is renamed (or saved with vim) (#1211)
Fixes #1208
Fixes #1157

The `file` might be `undefined` if it's a temporary file that's deleted before the transformers run.

The watcher close function gets called with an undefined file sometimes when you rename/delete a file. The default Node.js fs.watch function is known for having some bugs, but adding an undefined check before deleting the file fixes the problem until the webpack team fixes it.